+++ frontend/src/app/pages/exercises/exercise/shared/simulated-region-overview/radiogram-list/human-readable-radiogram-type.pipe.ts
@@ -10,6 +10,7 @@ const map: { [Key in ExerciseRadiogram['type']]: string } = {
     treatmentStatusRadiogram: 'Behandlungsstatus',
     vehicleCountRadiogram: 'Anzahl an Fahrzeugen',
     resourceRequestRadiogram: 'Anfrage nach Fahrzeugen',
+    transferCountsRadiogram: 'Anzahl abtransportierter Patienten',
 };
 
 @Pipe({