+++ shared/src/models/radiogram/exercise-radiogram.ts
@@ -8,6 +8,7 @@ import { Radiogram } from './radiogram';
 import { TreatmentStatusRadiogram } from './treatment-status-radiogram';
 import { VehicleCountRadiogram } from './vehicle-count-radiogram';
 import { ResourceRequestRadiogram } from './resource-request-radiogram';
+import { TransferCountsRadiogram } from './transfer-counts-radiogram';
 
 export const radiograms = {
     MaterialCountRadiogram,
@@ -15,6 +16,7 @@ export const radiograms = {
     PatientCountRadiogram,
     PersonnelCountRadiogram,
     ResourceRequestRadiogram,
+    TransferCountsRadiogram,
     TreatmentStatusRadiogram,
     VehicleCountRadiogram,
 };
@@ -33,6 +35,7 @@ export const radiogramDictionary: ExerciseRadiogramDictionary = {
     patientCountRadiogram: PatientCountRadiogram,
     personnelCountRadiogram: PersonnelCountRadiogram,
     resourceRequestRadiogram: ResourceRequestRadiogram,
+    transferCountsRadiogram: TransferCountsRadiogram,
     treatmentStatusRadiogram: TreatmentStatusRadiogram,
     vehicleCountRadiogram: VehicleCountRadiogram,
 };